[python] remove messages_tool.py
